- Added AnglesUtils for common methods - moved startBuild from Log4j2 appender to the AbstractAnglesTestCase
